+++ test/derailed_benchmarks/require_tree_test.rb
@@ -9,7 +9,7 @@ class RequireTree < ActiveSupport::TestCase
   end
 
   def teardown
-    DerailedBenchmarks::RequireTree.const_set("REQUIRED_BY",  {})
+    DerailedBenchmarks::RequireTree.reset!
   end
 
   test "default_cost" do